# BUG/ENH: Use pyarrow grouped aggregation functions for pyarrow-backed groupby ops