- [x] closes #54627
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

Use pyarrow's `TableGroupBy` functionality for pyarrow-backed groupby aggregations.

This has a few benefits over the current approach of round-tripping through non-arrow dtypes:
1. proper support for groupby operations with pyarrrow decimal types
2. removes values-dependent behavior (e.g. https://github.com/pandas-dev/pandas/issues/54627#issuecomment-1684160350)
3. aligns grouped aggregation behavior with non-grouped aggregation behavior (existing implementation has a number of subtle differences)

Performance impact is mixed. I think the consistency benefits mentioned above outweigh the performance impact here. If others disagree, we could continue using the pandas implementation for specific aggregations. 

Note: These timings are from a slow laptop. It would be great if someone could run this ASV to confirm (@rhshadrach possibly?)

`asv continuous -f 1.1 upstream/main arrow-groupby -b groupby.GroupByAggregateArrowDtypes`

```
           before            after    ratio
+      32.3±0.9ms         54.8±3ms     1.70  groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'all')
+        35.3±1ms         51.4±2ms     1.46  groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'any')
+        36.1±1ms       50.2±0.9ms     1.39  groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'all')
+        39.8±6ms         54.6±2ms     1.37  groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'any')
+        42.6±2ms         57.1±3ms     1.34  groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'min')
+      42.5±0.9ms         54.2±6ms     1.28  groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'min')
+        40.7±3ms         51.4±3ms     1.26  groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'max')
+        40.9±2ms         50.8±4ms     1.24  groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'max')
+        46.2±3ms         56.4±2ms     1.22  groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'std')
+        41.4±2ms         49.3±1ms     1.19  groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'std')
+        43.3±2ms         50.7±1ms     1.17  groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'min')
-        427±80ms          128±4ms     0.30  groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'prod')
-        395±40ms         91.3±6ms     0.23  groupby.GroupByAggregateArrowDtypes.time_frame_agg(string[pyarrow], 'max')
-         401±9ms         89.3±3ms     0.22  groupby.GroupByAggregateArrowDtypes.time_frame_agg(string[pyarrow], 'min')
-        429±50ms         89.0±8ms     0.21  groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'max')
-        449±80ms         89.1±2ms     0.20  groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'min')
-        384±60ms         56.1±2ms     0.15  groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'sum')
-        442±80ms         56.6±2ms     0.13  groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'mean')
```


Longer term, there are potentially big performance gains if all arrow-backed aggregations can be passed to pyarrow at once. At the moment, we're iterating column by column. Here is a small example showing the impact:

```
import pandas as pd
import numpy as np 
import pyarrow as pa

N = 100_000
columns = list("abcdefghij")

data = np.random.randn(N, len(columns))
df = pd.DataFrame(data, columns=columns, dtype="float64[pyarrow]")
df["key"] = np.random.randint(0, 100, N)
table = pa.table(df)

# column-by-column:
%timeit [table.group_by("key").aggregate([(col, "sum")]) for col in columns]

# all at once:
%timeit table.group_by("key").aggregate([(col, "sum") for col in columns])

# 24.7 ms ± 1.17 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  <- column-by-column
# 4.54 ms ± 152 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <- all at once
```