@rhshadrach 2023-09-29
> I'm going to rerun my benchmarks in the next day or so to be sure.

Same results approximately. 

<details>

```
| Change   | Before [824a2738] <main>   | After [adb0d321] <arrow-groupby>   |   Ratio | Benchmark (Parameter)                                                                     |
|----------|----------------------------|------------------------------------|---------|-------------------------------------------------------------------------------------------|
| +        | 9.03±0.2ms                 | 29.2±3ms                           |    3.24 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'any')          |
| +        | 9.28±0.4ms                 | 28.8±1ms                           |    3.1  | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'max')          |
| +        | 8.83±0.1ms                 | 26.1±0.7ms                         |    2.95 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'min')  |
| +        | 8.31±0.07ms                | 24.4±2ms                           |    2.93 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'min')                 |
| +        | 7.78±0.1ms                 | 22.7±0.2ms                         |    2.92 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'all')                 |
| +        | 8.86±0.09ms                | 24.8±1ms                           |    2.8  | groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'max')  |
| +        | 9.27±0.3ms                 | 25.9±3ms                           |    2.79 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'min')          |
| +        | 9.08±0.03ms                | 25.0±2ms                           |    2.75 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'any')                 |
| +        | 8.13±0.09ms                | 22.3±0.1ms                         |    2.74 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'min')                |
| +        | 8.31±0.04ms                | 22.2±0.2ms                         |    2.67 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'max')                |
| +        | 8.66±0.2ms                 | 23.1±2ms                           |    2.67 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'all')          |
| +        | 10.9±0.2ms                 | 29.1±3ms                           |    2.66 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'max')                  |
| +        | 8.21±0.1ms                 | 21.2±0.8ms                         |    2.58 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'prod')               |
| +        | 8.82±0.2ms                 | 22.8±2ms                           |    2.58 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'max')                 |
| +        | 10.4±0.4ms                 | 26.0±3ms                           |    2.51 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'sum')          |
| +        | 8.81±0.08ms                | 21.7±0.2ms                         |    2.46 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'all')                |
| +        | 10.9±0.06ms                | 26.6±5ms                           |    2.44 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'min')                  |
| +        | 9.60±0.4ms                 | 22.3±2ms                           |    2.32 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'mean')                |
| +        | 10.2±0.1ms                 | 23.3±0.2ms                         |    2.28 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'std')                |
| +        | 9.94±0.1ms                 | 22.2±1ms                           |    2.23 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'any')                |
| +        | 11.9±0.03ms                | 26.2±3ms                           |    2.19 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'std')                  |
| +        | 12.4±0.1ms                 | 27.2±2ms                           |    2.19 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'std')  |
| +        | 9.52±0.2ms                 | 20.7±0.2ms                         |    2.18 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'mean')               |
| +        | 9.79±0.03ms                | 21.1±2ms                           |    2.16 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(double[pyarrow], 'sum')                |
| +        | 10.2±0.2ms                 | 22.1±0.2ms                         |    2.16 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'std')                 |
| +        | 12.3±0.5ms                 | 26.4±2ms                           |    2.15 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'mean')         |
| +        | 10.2±1ms                   | 21.2±0.9ms                         |    2.09 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'prod')                |
| +        | 10.7±0.09ms                | 22.4±2ms                           |    2.09 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(int32[pyarrow], 'sum')                 |
| +        | 11.0±0.2ms                 | 22.8±1ms                           |    2.06 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'mean')                 |
| +        | 12.6±0.4ms                 | 25.0±0.2ms                         |    1.99 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(duration[ms][pyarrow], 'std')          |
| +        | 11.6±0.2ms                 | 22.8±2ms                           |    1.97 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(timestamp[s, tz=UTC][pyarrow], 'mean') |
| +        | 10.5±0.1ms                 | 20.7±1ms                           |    1.96 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'prod')                 |
| +        | 11.8±0.06ms                | 21.3±0.5ms                         |    1.81 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'sum')                  |
| +        | 14.6±0.2ms                 | 24.3±2ms                           |    1.67 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'any')                  |
| +        | 14.9±0.1ms                 | 24.3±3ms                           |    1.63 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(bool[pyarrow], 'all')                  |
| -        | 122±2ms                    | 85.1±10ms                          |    0.7  | groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'prod')    |
| -        | 94.6±0.6ms                 | 48.8±0.4ms                         |    0.52 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'max')     |
| -        | 93.4±1ms                   | 48.5±0.3ms                         |    0.52 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'min')     |
| -        | 80.3±0.9ms                 | 36.7±0.4ms                         |    0.46 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'sum')     |
| -        | 106±1ms                    | 45.6±5ms                           |    0.43 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(string[pyarrow], 'max')                |
| -        | 88.1±5ms                   | 37.0±0.09ms                        |    0.42 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(decimal128(25, 3)[pyarrow], 'mean')    |
| -        | 108±2ms                    | 45.6±5ms                           |    0.42 | groupby.GroupByAggregateArrowDtypes.time_frame_agg(string[pyarrow], 'min')                |
```

</details>